+++ CHANGELOG.md
@@ -12,6 +12,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   dependency set defined in the metaproject.
   [#97](https://github.com/amperity/lein-monolith/pull/97)
 
+### Changed
+- Dependency sets declared in a subproject profile.clj will override the project
+  `:managed-dependencies` rather than concatenating with the metaproject
+  `:managed-dependencies`.
+
 
 ## [1.9.1] - 2024-04-17