+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Dependency sets declared in a subproject profile.clj will override the project
   `:managed-dependencies` rather than concatenating with the metaproject
   `:managed-dependencies`.
+  [#97](https://github.com/amperity/lein-monolith/pull/97)
 
 
 ## [1.9.1] - 2024-04-17